fix: app only methods moved too app service
